factor out validation code into new package
